UPDATE COMPONENT MISSION
