+++ .stylelintrc.json
@@ -13,8 +13,8 @@
       {
         "ignoreAtRules": ["tailwind", "apply", "variants", "responsive", "screen"]
       }
-    ]
+    ],
+    "csstree/validator": true
   },
-  "csstree/validator": true,
   "ignoreFiles": ["build/**", "dist/**", "**/reset*.css", "**/bootstrap*.css", "**/*.js", "**/*.jsx"]
 }
